Relaxed position duration in the model
Added the idleTime attribute to the Execute goal tokens. I n order to
have a better model I also allowed the Pos token to last for longer
than a tick (except for now for the initial one) to allow the
representation of a vehicle staying at its end location for several
ticks.