corrected error for action Execution in auv model
Previoud model allowed the action of executing an auv.Execute to be
execute by another auv. This resulted on plans where the model computed
the action impact as if it was done by another auv.
This new model now properly enforce that a Task is “executed” by the
correct vehicle.