MRA/Sidescan: added support for Normalization (sic) and TVG Gain (sic) to ImcSidescanParser. These parameters are absolute misnomers and should be replaced by something meaningful in the future. Neutral parameters are:
* TVG Gain = 0.5
* Normalization = 127.5;

For now this is required for the BluePrint Subsea StarFish sidescan and
'TVG Gain' is used as 'Gain' and Normalization as 'Contrast'. From the
StarFish SDK Programmers Guide:

Having chosen a palette, the simplest approach is to directly map the
sonar data to the palette index, as both have 256 levels. However, this
leads to a display that appears very washed-out or over-exposed, as the
sonar data very rarely occupies the entire 127dB span available to it. To
make better use of the palette, two software implemented controls are
recommended - "Gain" and "Contrast".:

* Gain: the purpose of the gain control is to apply an offset (also in
  Decibels) to the sonar signal, effectively allowing it to be shifted
  across the palette range. As the sonar signal is also in Decibels, the
  gain value can be added to each signal value.

* Contrast: the purpose of the contrast control is to define the span (or
  width) of the colour palette in Decibels. In other words, if the black
  colour index (0) of the palette represented an arbitrary 0dB level, then
  a contrast of 40dB would mean that the palette index 255 would be used
  to represent a signal of 40dB, with linear mapping of all in-between
  palette index's to signal levels. Changing the contrast effectively
  stretches or compresses the sonar signal against the palette colours